Moved MockSystem to its own file.
